[fix] Fail loudly on empty DP device discovery; skip unsupported MPI tests.
