+++ tests/test_dp_modes.py
@@ -1,9 +1,13 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: 2025 Tenstorrent USA, Inc.
 
-"""Unit tests for TT standard-DP routing and launch semantics."""
+"""Unit tests for how ``check_and_update_config`` routes a DP config.
 
-import pathlib
+Engine-core/launcher/scheduler selection and the worker's device lifecycle. The
+``TT_VISIBLE_DEVICES`` group lifecycle lives in ``test_visible_devices.py``, and
+the DP-to-lanes conversion in ``test_galaxy_dp_conversion.py``.
+"""
+
 from types import SimpleNamespace
 
 import pytest
@@ -304,260 +308,3 @@ class TestDPModes:
                 match="TT standard DP does not support MoE models yet",
             ):
                 TTPlatform.check_and_update_config(vllm_config)
-
-    @pytest.mark.xfail(
-        reason="`vllm_tt_plugin.launcher` imports legacy vllm core-engine objects."
-    )
-    def test_standard_dp_uses_all_device_ranks(
-        self,
-        tmp_path: pathlib.Path,
-        vllm_config: SimpleNamespace,
-    ) -> None:
-        from vllm_tt_plugin.launcher import parse_tt_mpi_params
-
-        rank_binding = tmp_path / "rank_binding.json"
-        rank_binding.write_text(
-            "rank_bindings:\n"
-            "  - rank: 0\n"
-            "    mesh_id: 0\n"
-            "    env_overrides:\n"
-            '      TT_VISIBLE_DEVICES: "0"\n'
-            "\n"
-            "  - rank: 1\n"
-            "    mesh_id: 1\n"
-            "    env_overrides:\n"
-            '      TT_VISIBLE_DEVICES: "3"\n'
-            "\n"
-            "  - rank: 2\n"
-            "    mesh_id: 2\n"
-            "    env_overrides:\n"
-            '      TT_VISIBLE_DEVICES: "1"\n'
-            "\n"
-            "  - rank: 3\n"
-            "    mesh_id: 3\n"
-            "    env_overrides:\n"
-            '      TT_VISIBLE_DEVICES: "2"\n'
-        )
-
-        vllm_config.additional_config = {"tt": {"rank_binding": str(rank_binding)}}
-        vllm_config.parallel_config.data_parallel_backend = "mp"
-        vllm_config.parallel_config.data_parallel_size = 4
-
-        parsed_rank_binding, non_device_dp_ranks = parse_tt_mpi_params(vllm_config)
-
-        assert parsed_rank_binding == str(rank_binding)
-        assert non_device_dp_ranks == set()
-
-    @pytest.mark.xfail(
-        reason="`vllm_tt_plugin.launcher` imports legacy vllm core-engine objects."
-    )
-    def test_standard_dp_rejects_mismatched_mpi_world(
-        self,
-        tmp_path: pathlib.Path,
-        vllm_config: SimpleNamespace,
-    ) -> None:
-        from vllm_tt_plugin.launcher import parse_tt_mpi_params
-
-        rank_binding = tmp_path / "rank_binding.json"
-        rank_binding.write_text(
-            "rank_bindings:\n"
-            "  - rank: 0\n"
-            "    mesh_id: 0\n"
-            "    env_overrides:\n"
-            '      TT_VISIBLE_DEVICES: "0, 1"\n'
-            "\n"
-            "  - rank: 1\n"
-            "    mesh_id: 1\n"
-            "    env_overrides:\n"
-            '      TT_VISIBLE_DEVICES: "2, 3"\n'
-        )
-
-        vllm_config.additional_config = {"tt": {"rank_binding": str(rank_binding)}}
-        vllm_config.parallel_config.data_parallel_backend = "mp"
-        vllm_config.parallel_config.data_parallel_size = 4
-
-        with pytest.raises(
-            RuntimeError,
-            match="Standard DP mode requires one TT MPI rank per DP rank",
-        ):
-            parse_tt_mpi_params(vllm_config)
-
-    @pytest.mark.xfail(
-        reason="`vllm_tt_plugin.launcher` imports legacy vllm core-engine objects."
-    )
-    def test_explicit_mpi_args_require_rank_binding(
-        self,
-        vllm_config: SimpleNamespace,
-    ) -> None:
-        from vllm_tt_plugin.launcher import parse_tt_mpi_params
-
-        vllm_config.additional_config = {"tt": {"mpi_args": "--host hostA"}}
-        vllm_config.parallel_config.data_parallel_backend = "mp"
-        vllm_config.parallel_config.data_parallel_size = 4
-
-        with pytest.raises(
-            RuntimeError,
-            match="TT explicit MPI launch requires tt.rank_binding",
-        ):
-            parse_tt_mpi_params(vllm_config)
-
-    @pytest.mark.xfail(
-        reason="`vllm_tt_plugin.launcher` imports legacy vllm core-engine objects."
-    )
-    def test_multinode_requires_rank_binding(
-        self,
-        vllm_config: SimpleNamespace,
-    ) -> None:
-        from vllm_tt_plugin.launcher import parse_tt_mpi_params
-
-        vllm_config.additional_config = {"tt": {}}
-        vllm_config.parallel_config.data_parallel_backend = "mp"
-        vllm_config.parallel_config.data_parallel_size = 4
-        vllm_config.parallel_config.nnodes = 2
-
-        with pytest.raises(
-            RuntimeError,
-            match="TT explicit MPI launch requires tt.rank_binding",
-        ):
-            parse_tt_mpi_params(vllm_config)
-
-    @pytest.mark.xfail(
-        reason="`vllm_tt_plugin.launcher` imports legacy vllm core-engine objects."
-    )
-    def test_rank_binding_requires_visible_devices(
-        self,
-        tmp_path: pathlib.Path,
-        vllm_config: SimpleNamespace,
-    ) -> None:
-        from vllm_tt_plugin.launcher import parse_tt_mpi_params
-
-        rank_binding = tmp_path / "rank_binding.json"
-        rank_binding.write_text(
-            "rank_bindings:\n"
-            "  - rank: 0\n"
-            "    mesh_id: 0\n"
-            "    env_overrides:\n"
-            "      OTHER_ENV: foo\n"
-            "  - rank: 1\n"
-            "    mesh_id: 1\n"
-            "    env_overrides:\n"
-            '      TT_VISIBLE_DEVICES: "1"\n'
-        )
-
-        vllm_config.additional_config = {"tt": {"rank_binding": str(rank_binding)}}
-        vllm_config.parallel_config.data_parallel_backend = "mp"
-        vllm_config.parallel_config.data_parallel_size = 2
-
-        with pytest.raises(RuntimeError, match="TT_VISIBLE_DEVICES"):
-            parse_tt_mpi_params(vllm_config)
-
-    @pytest.mark.xfail(
-        reason="`vllm_tt_plugin.launcher` imports legacy vllm core-engine objects."
-    )
-    def test_rank_binding_rejects_overlapping_visible_devices(
-        self,
-        tmp_path: pathlib.Path,
-        vllm_config: SimpleNamespace,
-    ) -> None:
-        from vllm_tt_plugin.launcher import parse_tt_mpi_params
-
-        rank_binding = tmp_path / "rank_binding.json"
-        rank_binding.write_text(
-            "rank_bindings:\n"
-            "  - rank: 0\n"
-            "    mesh_id: 0\n"
-            "    env_overrides:\n"
-            '      TT_VISIBLE_DEVICES: "0, 1"\n'
-            "  - rank: 1\n"
-            "    mesh_id: 1\n"
-            "    env_overrides:\n"
-            '      TT_VISIBLE_DEVICES: "1, 2"\n'
-        )
-
-        vllm_config.additional_config = {"tt": {"rank_binding": str(rank_binding)}}
-        vllm_config.parallel_config.data_parallel_backend = "mp"
-        vllm_config.parallel_config.data_parallel_size = 2
-
-        with pytest.raises(
-            RuntimeError,
-            match="overlaps TT_VISIBLE_DEVICES assignments",
-        ):
-            parse_tt_mpi_params(vllm_config)
-
-    @pytest.mark.xfail(
-        reason="`vllm_tt_plugin.launcher` imports legacy vllm core-engine objects."
-    )
-    def test_rank_binding_rejects_duplicate_rank_ids(
-        self,
-        tmp_path: pathlib.Path,
-        vllm_config: SimpleNamespace,
-    ) -> None:
-        from vllm_tt_plugin.launcher import parse_tt_mpi_params
-
-        rank_binding = tmp_path / "rank_binding.json"
-        rank_binding.write_text(
-            "rank_bindings:\n"
-            "  - rank: 0\n"
-            "    mesh_id: 0\n"
-            "    env_overrides:\n"
-            '      TT_VISIBLE_DEVICES: "0"\n'
-            "  - rank: 0\n"
-            "    mesh_id: 1\n"
-            "    env_overrides:\n"
-            '      TT_VISIBLE_DEVICES: "1"\n'
-        )
-
-        vllm_config.additional_config = {"tt": {"rank_binding": str(rank_binding)}}
-        vllm_config.parallel_config.data_parallel_backend = "mp"
-        vllm_config.parallel_config.data_parallel_size = 2
-
-        with pytest.raises(RuntimeError, match="duplicate rank 0"):
-            parse_tt_mpi_params(vllm_config)
-
-    @pytest.mark.xfail(
-        reason="`vllm_tt_plugin.launcher` imports legacy vllm core-engine objects."
-    )
-    def test_legacy_gathered_override_is_ignored_by_launcher(
-        self,
-        tmp_path: pathlib.Path,
-        vllm_config: SimpleNamespace,
-    ) -> None:
-        from vllm_tt_plugin.launcher import parse_tt_mpi_params
-
-        rank_binding = tmp_path / "rank_binding.json"
-        rank_binding.write_text(
-            "rank_bindings:\n"
-            "  - rank: 0\n"
-            "    mesh_id: 0\n"
-            "    env_overrides:\n"
-            '      TT_VISIBLE_DEVICES: "0"\n'
-            "\n"
-            "  - rank: 1\n"
-            "    mesh_id: 1\n"
-            "    env_overrides:\n"
-            '      TT_VISIBLE_DEVICES: "1"\n'
-            "\n"
-            "  - rank: 2\n"
-            "    mesh_id: 2\n"
-            "    env_overrides:\n"
-            '      TT_VISIBLE_DEVICES: "2"\n'
-            "\n"
-            "  - rank: 3\n"
-            "    mesh_id: 3\n"
-            "    env_overrides:\n"
-            '      TT_VISIBLE_DEVICES: "3"\n'
-        )
-
-        vllm_config.additional_config = {
-            "tt": {
-                "rank_binding": str(rank_binding),
-                "tt_data_parallel_size": 4,
-            }
-        }
-        vllm_config.parallel_config.data_parallel_backend = "mp"
-        vllm_config.parallel_config.data_parallel_size = 4
-
-        parsed_rank_binding, non_device_dp_ranks = parse_tt_mpi_params(vllm_config)
-
-        assert parsed_rank_binding == str(rank_binding)
-        assert non_device_dp_ranks == set()
